requirements.txt
typer>=0.9
PyYAML>=6.0
pydantic>=2.0
